requirements/base.txt
#Django
Django==4.1.3

# DRF
djangorestframework==3.14.0
drf-yasg==1.21.4
django-filter==22.1
drf-spectacular[sidecar]==0.24.2
djangorestframework-simplejwt==5.2.2

# Celery
celery==5.2.7
flower==1.2.0
django-celery-beat==2.4.0
django-celery-results==2.4.0
tenant-schemas-celery==2.0.0

# DB
psycopg2-binary==2.9.5

# Passwords security
argon2-cffi==21.3.0

# Firebase
firebase-admin==6.0.1

# Channels
channels[daphne]==4.0.0
channels-redis==4.0.0

# Django Tenants
django-tenants==3.4.7

# JWT
pyjwt==2.6.0

# Utils
Pillow
ipython
python-dotenv==0.21.0
wheel==0.38.4

# CORS
django-cors-headers==3.13.0

# Countries
django-countries==7.4.2

# Currency
currencies==2020.12.12

# OTP
pyotp==2.7.0

# Twilio
twilio==7.16.0

# IPFS
ipfshttpclient==0.8.0a1

# PDF
PyPDF2==3.0.1

# Stripe
stripe==5.4.0

# Pycryptodome
pycryptodome==3.17

# Pinata
pinata==0.0.1
